pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ccx-collab"
version = "0.4.0"
description = "CLI tool for Claude Code + Codex CLI collaboration pipeline"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "jsonschema>=4.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "jinja2>=3.1",
    "httpx>=0.25",
    "aiosqlite>=0.19",
    "python-multipart>=0.0.6",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pre-commit>=3.0",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://github.com/tornado1014/ccx-collab"
Repository = "https://github.com/tornado1014/ccx-collab"
Issues = "https://github.com/tornado1014/ccx-collab/issues"

[project.scripts]
ccx-collab = "ccx_collab.cli:main"

[tool.pytest.ini_options]
markers = ["windows: Windows-specific tests (deselect with '-m not windows')"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["ccx_collab*"]
